connect-examples/oauth/python/requirements.txt
CacheControl==0.12.6
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
deprecation==2.1.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==0.9.6
MarkupSafe==1.1.1
msgpack==1.0.0
packaging==20.3
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
squareup==5.2.2.20200422
urllib3==1.25.9
Werkzeug==1.0.1
